requirements.txt
discord.py>=2.3.0
playwright>=1.40.0
asyncio
discord.py
python-dotenv
playwright